+++ CheckInSystem/Models/Absence.cs
@@ -31,9 +31,10 @@ namespace CheckInSystem.Models
         public DateTime FromDate { get; set; }
         public DateTime ToDate { get; set; }
         public absenceReason AbsenceReason { get; set; } 
-        public string? Note { get; set; } 
+        public string? Note { get; set; }
 
-        
+        private TimeSpan waitTime = TimeSpan.FromMinutes(5);
+
         public Absence(Absence absence)
         {
             ID = absence.ID;
@@ -96,9 +97,66 @@ namespace CheckInSystem.Models
             }
         }
 
-        public void OffsiteTimer()
+        public async Task OffsiteTimer(Employee employee)
         {
+            while (true == true)
+            {
+                var activeAbsences = Absence.GetAllAbsence(employee)
+                        .Where(a => a.ToDate > DateTime.Now).ToList();
 
+                if (activeAbsences == null)
+                    return;
+
+                foreach (var activeAbsence in activeAbsences)
+                {
+                    // Wait until FromDate while checking every 5 minutes if the absence still exists
+                    while (activeAbsence.FromDate > DateTime.Now)
+                    {
+                        var CurrentAbsences = Absence.GetAllAbsence(employee)
+                            .Where(a => a.ID == activeAbsence.ID);
+
+                        foreach (var absence in CurrentAbsences)
+                            activeAbsence.FromDate = absence.FromDate;
+
+                        if (!CurrentAbsences.Any(a => a.ID == activeAbsence.ID))
+                            return;
+
+                        var timeUntilStart = activeAbsence.FromDate - DateTime.Now;
+
+                        //it uses whatever time is shortest for the constent checking if the absence still exist
+                        await Task.Delay(timeUntilStart < waitTime ? timeUntilStart : waitTime);
+                    }
+
+                    employee.IsOffSite = true;
+
+                    // Wait until ToDate while checking every 5 minutes if the absence still exists
+                    while (activeAbsence.ToDate > DateTime.Now)
+                    {
+                        var CurrentAbsences = Absence.GetAllAbsence(employee)
+                            .Where(a => a.ID == activeAbsence.ID);
+
+                        foreach (var absence in CurrentAbsences)
+                            activeAbsence.ToDate = absence.ToDate;
+
+                        if (!CurrentAbsences.Any(a => a.ID == activeAbsence.ID))
+                        {
+                            employee.IsOffSite = false;
+
+                            return;
+                        }
+
+                        var timeUntilEnd = activeAbsence.ToDate - DateTime.Now;
+
+                        //it uses whatever time is shortest for the constent checking if the absence still exist
+                        await Task.Delay(timeUntilEnd < waitTime ? timeUntilEnd : waitTime);
+
+                    }
+
+                    employee.IsOffSite = false;
+
+                }
+                await Task.Delay(waitTime);
+            }
         }
         public Absence()
         {
